Moved the utils/extprocess.py to the cbooks module - deleted unused import in cbooks
