+++ lib/matplotlib/backends/backend_pgf.py
@@ -19,7 +19,8 @@ from matplotlib import _png, rcParams
 from matplotlib import font_manager
 from matplotlib.ft2font import FT2Font
 from matplotlib.cbook import is_string_like, is_writable_file_like
-from matplotlib.utils.extprocess import check_output
+from matplotlib.cbook import check_output
+
 
 ###############################################################################
