+++ lib/matplotlib/cbook.py
@@ -12,8 +12,9 @@ import numpy.ma as ma
 from weakref import ref, WeakKeyDictionary
 import cPickle
 import os.path
-import random
 from functools import reduce
+
+import subprocess
 
 import matplotlib
 
@@ -1895,6 +1896,48 @@ except AttributeError:
 else:
     def _putmask(a, mask, values):
         return np.copyto(a, values, where=mask)
+
+
+def _check_output(*popenargs, **kwargs):
+    r"""Run command with arguments and return its output as a byte
+    string.
+
+    If the exit code was non-zero it raises a CalledProcessError.  The
+    CalledProcessError object will have the return code in the
+    returncode
+    attribute and output in the output attribute.
+
+    The arguments are the same as for the Popen constructor.  Example:
+
+    >>> check_output(["ls", "-l", "/dev/null"])
+    'crw-rw-rw- 1 root root 1, 3 Oct 18  2007 /dev/null\n'
+
+    The stdout argument is not allowed as it is used internally.
+    To capture standard error in the result, use stderr=STDOUT.
+
+    >>> check_output(["/bin/sh", "-c",
+    ...               "ls -l non_existent_file ; exit 0"],
+    ...              stderr=STDOUT)
+    'ls: non_existent_file: No such file or directory\n'
+    """
+    if 'stdout' in kwargs:
+        raise ValueError('stdout argument not allowed, it will be overridden.')
+    process = subprocess.Popen(stdout=subprocess.PIPE, *popenargs, **kwargs)
+    output, unused_err = process.communicate()
+    retcode = process.poll()
+    if retcode:
+        cmd = kwargs.get("args")
+        if cmd is None:
+            cmd = popenargs[0]
+        raise subprocess.CalledProcessError(retcode, cmd, output=output)
+    return output
+
+
+# python2.7's subprocess provides a check_output method
+if hasattr(subprocess, 'check_output'):
+    check_output = subprocess.check_output
+else:
+    check_output = _check_output
 
 
 if __name__=='__main__':
